Filtering
Added a filtering that I don't really know how it works yet but it kinda does. Will needed to add more with the other filters and need to add a bubble or the tree map to finish off all of our visualizations so we can move onto other things such as hovering/click filtering along with the dropdown. Potentially change the dropdown so we can also select multiple artists.