Almost Done
Everything is close to being done, may add an additional feature of a textbox to allow the user to filter for the top "x" streams. Sorted the bubble chart so smaller bubbles show on top always, fixed a few bugs and added more filtering so only data that is valid will show up. Fixed loudness so that it will show all of the negative numbers.