Final Version
Fixed a few things, finally figure out why I didn't have axis lines and fixed that, changed it back so that the bubble chart isn't sorted by size because it messed with the color but I couldn't figure out how to sort it by size and color it by another way. Fixed some of the css.